requirements.txt
# Core dependencies
torch>=1.10.0
numpy>=1.20.0
ray>=2.0.0
psutil>=5.9.0
python-chess>=1.9.0

# Visualization dependencies
matplotlib>=3.5.0

# Optional - for monitoring
tensorboard>=2.8.0
pyyaml>=6.0

# Development dependencies
pytest>=7.0.0 